Corrections in background processes

+++ src/odcommon/usr/local/opendomo/services/syscript/saveSystemConfig.sh
@@ -6,17 +6,23 @@
 SYSUPDATEPID="/var/opendomo/run/updateSystem.pid"
 PIDFILE="/var/opendomo/run/saveSystemConf.pid"
 
-# creating PID
+# Check process
+if test -f $PIDFILE; then
+    echo "#ERRO Save config is already running"
+    exit 1
+fi
+
+# Creating PID
 touch $PIDFILE
 
 # Waiting system update in first configuration script
-while -f $SYSUPDATEPID; do
+while test -f $SYSUPDATEPID; do
     sleep 1
 done
 
-echo -n "#INFO Save system configuration (please wait) ..."
+echo "#LOADING Save system configuration (please wait) ..."
+echo
 sudo /usr/local/sbin/mkrootfs custom >/dev/null 2>/dev/null
-echo "   (done)"
 rm $PIDFILE
 
 /usr/local/opendomo/manageConfiguration.sh
